requirements.txt
chardet==3.0.4
pandas==0.25.1
pyyaml==5.1.2